% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peekds_readers.R
\name{ds.is_table_required}
\alias{ds.is_table_required}
\title{Check if a certain table is required according to schema}
\usage{
ds.is_table_required(table_type, coding_methods)
}
\arguments{
\item{table_type}{the type of dataframe, for the most updated table types
specified by schema, please use ds.list_ds_tables()}

\item{coding_methods}{methods used in the experiment for coding gaze data, to
 get the list of current coding methods, please use function
ds.list_coding_methods()}
}
\value{
A boolean value
}
\description{
Check if a certain table is required according to schema
}
\examples{
\dontrun{
is_required <- ds.is_table_required(table_type = "xy_timepoints",
                                 coding_method = "manual gaze coding")
}
}
