% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peekds_readers.R
\name{ds.list_coding_methods}
\alias{ds.list_coding_methods}
\title{Get the coding method list from json schema file}
\usage{
ds.list_coding_methods()
}
\value{
a list of strings indicating allowed coding methods
}
\description{
Get the coding method list from json schema file
}
\examples{
\dontrun{
coding_methods <- ds.list_coding_methods()
}
}
