% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peekds_readers.R
\name{ds.list_ds_tables}
\alias{ds.list_ds_tables}
\title{List the tables required based on coding method}
\usage{
ds.list_ds_tables(coding_methods = c("eyetracking"))
}
\arguments{
\item{coding_methods}{a list of strings indicating the methods used in the experiment for
coding gaze data, to get the list of current coding methods, please use
ds.list_coding_methods()}
}
\value{
a list of table types that are required based on input coding method
}
\description{
List the tables required based on coding method
}
\examples{
\dontrun{
table_list <- ds.list_ds_tables(coding_methods = "manual gaze coding")
}

}
