% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peekds_readers.R
\name{ds.list_language_choices}
\alias{ds.list_language_choices}
\title{List current allowed language choices for db import}
\usage{
ds.list_language_choices()
}
\value{
a list of strings containing all the allowed language codes based on
  json schema file
}
\description{
List current allowed language choices for db import
}
\examples{
\dontrun{
language_list <- ds.list_language_choices()
}

}
