% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peekds_readers.R
\name{ds.map_columns}
\alias{ds.map_columns}
\title{Function for mapping raw data columns to processed table columns}
\usage{
ds.map_columns(raw_data, raw_format, table_type)
}
\arguments{
\item{raw_data}{raw data frame}

\item{raw_format}{source of the eye-tracking data, e.g. "tobii"}

\item{table_type}{type of processed table, e.g. "xy_data" | "aoi_table"}
}
\value{
processed data frame with specified column names
}
\description{
Function for mapping raw data columns to processed table columns
}
\examples{
\dontrun{
df_xy_data <- ds.map_columns(raw_data = raw_data, raw_format = "tobii",
                          table_type = "xy_data")
df_aoi_data <- ds.map_columns(raw_data = raw_data, raw_format = "tobii",
                           table_type = "aoi_data")
}

}
