% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peekds_generate_aoi.R
\name{ds.normalize_times}
\alias{ds.normalize_times}
\title{sets the starting point of a given trial to be zero}
\usage{
ds.normalize_times(df_table)
}
\arguments{
\item{df_table}{to-be-resampled dataframe with t, aoi/xy values, trial_id and
administration_id}
}
\value{
df_out with resampled time, xy or aoi value rows
}
\description{
sets the starting point of a given trial to be zero
}
