% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peeksd_osf.R
\name{ds.put_processed_data}
\alias{ds.put_processed_data}
\title{Put processed data for specific peekbank dataset on OSF}
\usage{
ds.put_processed_data(token, dataset_name, path = ".", osf_address = "pr6wu")
}
\arguments{
\item{token}{personal access tokens for uploading to OSF}

\item{dataset_name}{Specific dataset name occurring in the file hierarchy of
the relevant OSF repo.}

\item{path}{Where the data live on your own machine.}

\item{osf_address}{pr6wu for peekbank.}
}
\description{
Put processed data for specific peekbank dataset on OSF
}
