% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peekds_validators.R
\name{file.exists.case.sensitive}
\alias{file.exists.case.sensitive}
\title{Check if a file exists with exact case sensitivity}
\usage{
file.exists.case.sensitive(...)
}
\arguments{
\item{...}{character vectors, containing file paths}
}
\value{
logical value: TRUE if the file exists with the exact same case,
  FALSE otherwise
}
\description{
Check if a file exists with exact case sensitivity
}
\examples{
\dontrun{
exists <- file.exists.case.sensitive("path/to/image.jpg")
}

}
