% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peekbankr.R
\name{get_aoi_data}
\alias{get_aoi_data}
\title{Get AOI data}
\usage{
get_aoi_data(connection = NULL)
}
\arguments{
\item{connection}{A connection to the peekbank database}
}
\value{
A `tbl` of AOI data, filtered down by supplied arguments. If
  `connection` is supplied, the result remains a remote query, otherwise it
  is retrieved into a local tibble.
}
\description{
Get AOI data
}
\examples{
get_aoi_data
}
