% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peekbankr.R
\name{get_aoi_regions}
\alias{get_aoi_regions}
\title{Get AOI regions}
\usage{
get_aoi_regions(connection = NULL)
}
\arguments{
\item{connection}{A connection to the peekbank database}
}
\value{
A `tbl` of AOI Regions data, filtered down by supplied arguments. If
  `connection` is supplied, the result remains a remote query, otherwise it
  is retrieved into a local tibble.
}
\description{
Get AOI regions
}
\examples{
get_aoi_regions()
}
