% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peekbankr.R
\name{get_datasets}
\alias{get_datasets}
\title{Get datasets}
\usage{
get_datasets(connection = NULL)
}
\arguments{
\item{connection}{A connection to the peekbank database}
}
\value{
A tibble of Datasets data.
}
\description{
Get datasets
}
\examples{
\dontrun{
get_datasets()
}
}
