% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peekbankr.R
\name{get_db_info}
\alias{get_db_info}
\title{Get information on database connection options}
\usage{
get_db_info()
}
\value{
List of database info: host name, current version, supported
  versions, historical versions, username, password
}
\description{
Get information on database connection options
}
\examples{
\dontrun{
get_db_info()
}
}
