% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peekbankr.R
\name{get_sql_query}
\alias{get_sql_query}
\title{Run a SQL query against peekbank}
\usage{
get_sql_query(sql_query_string, connection = NULL)
}
\arguments{
\item{sql_query_string}{A valid SQL query string}

\item{connection}{A connection to the peekbank database}
}
\value{
A tibble of the query results
}
\description{
Queries are BigQuery Standard SQL, run against the peekbank dataset on
Redivis. Note that string comparison is case-sensitive.
}
\examples{
\dontrun{
con <- connect_to_peekbank()
get_sql_query("SELECT * FROM datasets", connection = con)
}
}
