% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peekbankr.R
\name{list_peekbank_tables}
\alias{list_peekbank_tables}
\title{List of peekbank tables}
\usage{
list_peekbank_tables(connection)
}
\arguments{
\item{connection}{A connection to the peekbank database}
}
\value{
A vector of the names of tables in peekbank
}
\description{
List of peekbank tables
}
\examples{
\dontrun{
con <- connect_to_peekbank()
list_peekbank_tables(con)
}
}
