% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peekbankr.R
\name{unpack_aux_data}
\alias{unpack_aux_data}
\title{Unpack the json sting in the *_aux_data column and turns
it into a nested R list}
\usage{
unpack_aux_data(df)
}
\arguments{
\item{df}{a dataframe in the peekbank format that has an aux data column}
}
\value{
the input dataframe, with the *_aux_data column unpacked
}
\description{
Unpack the json sting in the *_aux_data column and turns
it into a nested R list
}
\examples{
\dontrun{
subjects_table <- unpack_aux_data(df = subjects_table)
}
}
