% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdi_processing.R
\name{append_relative_cdi_scores}
\alias{append_relative_cdi_scores}
\title{Adds a relative cdi score indicating the percentage of total achievable points the subject got on each given measure}
\usage{
append_relative_cdi_scores(subjects_table)
}
\arguments{
\item{subjects_table}{a subjects table with unnested cdi data, needs columns "subject_id", "language", "instrument_type", "measure", "rawscore"}
}
\value{
the input table with an added "cdi_relative" column that contains the percentage of total points gained in the given administrations
}
\description{
Adds a relative cdi score indicating the percentage of total achievable points the subject got on each given measure
}
\examples{
\donttest{
cdi_data <- all_subjects \%>\%
  unnest(subject_aux_data) \%>\%
  filter(!is.na(cdi_responses)) \%>\%
  unnest(cdi_responses) \%>\%
  append_relative_cdi_scores()
}
}
