% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdi_processing.R
\name{cleanup_cdi_data}
\alias{cleanup_cdi_data}
\title{Checks cdi data for inconsistencies, warns about them, and fixes them}
\usage{
cleanup_cdi_data(cdi_data)
}
\arguments{
\item{cdi_data}{a subjects table with unnested cdi data, needs columns "subject_id", "language", "instrument_type", "age", "sex", "measure", "rawscore"}
}
\value{
a cleaned up version of the cdi data
}
\description{
Checks cdi data for inconsistencies, warns about them, and fixes them
}
\examples{
\donttest{
clean_cdi_data <- all_subjects \%>\%
  unnest(subject_aux_data) \%>\%
  filter(!is.na(cdi_responses)) \%>\%
  unnest(cdi_responses) \%>\%
  peekbankr::cleanup_cdi_data()
}
}
