% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peekbankr.R
\name{connect_to_peekbank}
\alias{connect_to_peekbank}
\title{Connect to Peekbank}
\usage{
connect_to_peekbank(
  db_version = "current",
  db_args = NULL,
  compress = TRUE,
  host = NULL,
  port = NULL
)
}
\arguments{
\item{db_version}{String of the name of database version to use}

\item{db_args}{List with host, user, and password defined}

\item{compress}{Flag to use compression protocol (defaults to TRUE)}

\item{host}{Hostname of the Peekbank server to connect to (defaults hosted PB)}

\item{port}{Port of the Peekbank DB to connect to (defaults to 3306)}
}
\value{
con A DBIConnection object for the peekbank database
}
\description{
Connect to Peekbank
}
\examples{
\donttest{
con <- connect_to_peekbank(db_version = "current", db_args = NULL)
DBI::dbDisconnect(con)
}
}
