% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peekbankr.R
\name{connect_to_peekbank}
\alias{connect_to_peekbank}
\title{Connect to the peekbank database}
\usage{
connect_to_peekbank()
}
\value{
A connection to the peekbank database.
}
\description{
Connect to the peekbank database
}
\examples{
con <- connect_to_peekbank()
DBI::dbDisconnect(con)
}
