% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peekbankr.R
\name{connect_to_peekbank}
\alias{connect_to_peekbank}
\title{Connect to peekbank}
\usage{
connect_to_peekbank(db_version = "current", db_args = NULL)
}
\arguments{
\item{db_version}{String of the name of database version to use}

\item{db_args}{List with host, user, and password defined}
}
\value{
con A DBIConnection object for the peekbank database
}
\description{
Connect to peekbank
}
\examples{
\donttest{
con <- connect_to_peekbank(db_version = "current", db_args = NULL)
DBI::dbDisconnect(con)
}
}
