% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peekbankr.R
\name{download_osf_files}
\alias{download_osf_files}
\title{Download a list of files from OSF and recreate folder structure locally}
\usage{
download_osf_files(
  file_paths,
  osf_node_id = "pr6wu",
  local_base_dir = "data",
  debug = F,
  skip_existing = TRUE
)
}
\arguments{
\item{file_paths}{A character vector of file paths on OSF to download}

\item{osf_node_id}{The OSF node ID where the files are stored (default: "pr6wu")}

\item{local_base_dir}{Base directory to save files locally (default: here::here("data"))}

\item{debug}{Logical, whether to print debugging information (default: TRUE)}

\item{skip_existing}{Logical, skip downloading a file if a file with that name already exists in that path locally}
}
\value{
returns paths to downloaded files
}
\description{
Download a list of files from OSF and recreate folder structure locally
}
\examples{
\dontrun{
# Download multiple files from OSF
download_osf_files(
  file_paths = c(
    "lab1/raw_data/file1.csv",
    "lab2/processed_data/file2.csv"
  ),
  osf_node_id = "pr6wu"
)
}
}
