% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peekds_generate_aoi.R
\name{ds.compute_aois}
\alias{ds.compute_aois}
\title{Classify gaze samples into AOIs based x/y coordinates and provided region sets}
\usage{
ds.compute_aois(data)
}
\arguments{
\item{data}{a data frame carrying x, y, target_side, the eight region bounds
(l_x_max, l_x_min, l_y_max, l_y_min, r_x_max, r_x_min, r_y_max, r_y_min) and
monitor_size_x / monitor_size_y}
}
\value{
\code{data} with the \code{aoi} column set, overwriting any existing value
}
\description{
Classify gaze samples into AOIs based x/y coordinates and provided region sets
}
