% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peekds_readers.R
\name{ds.get_json_fields}
\alias{ds.get_json_fields}
\title{Fetching the list of field names and requirements in each table according to
the schema json file}
\usage{
ds.get_json_fields(table_type)
}
\arguments{
\item{table_type}{the type of dataframe, for the most updated table types
specified by schema, please use functionds.list_ds_tables()}
}
\value{
the list of field names
}
\description{
Fetching the list of field names and requirements in each table according to
the schema json file
}
\examples{
\dontrun{
fields_json <-ds.get_json_fields(table_type = "aoi_timepoints")
}

}
