% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peekds_readers.R
\name{ds.get_peekjson}
\alias{ds.get_peekjson}
\title{parse the bundled peekbank schema json into a dataframe}
\usage{
ds.get_peekjson()
}
\value{
the organized dataframe from schema json file
}
\description{
parse the bundled peekbank schema json into a dataframe
}
\examples{
\dontrun{
peekjson <-ds.get_peekjson()
}

}
