% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peekds_validators.R
\name{ds.validate_table}
\alias{ds.validate_table}
\title{Check if a dataframe/table is compliant to peekbank json before database
import}
\usage{
ds.validate_table(
  df_table,
  table_type,
  cdi_expected,
  dir_csv,
  is_null_field_required = TRUE
)
}
\arguments{
\item{df_table}{the dataframe to be saved}

\item{table_type}{the type of dataframe, for the most updated table types
specified by schema, please use functionds.list_ds_tables()}

\item{cdi_expected}{specifies whether cdi_data is to be expected to be
present in the imported data; only relevant for subjects table}

\item{dir_csv}{the folder directory containing all the csv files, used for
stimulus image path validation}

\item{is_null_field_required}{by default is set to TRUE which means that
all the columns in the json file are required; when user specifically
sets this to FALSE, then the fields that are allowed null values are not
required.}
}
\value{
A list with two elements:
  \describe{
    \item{errors}{Character vector of validation errors (blocking), or NULL if none.}
    \item{warnings}{Named list of validation warnings (non-blocking). Names are warning IDs
      (e.g. \code{"duplicate_cdi"}).}
  }
}
\description{
Check if a dataframe/table is compliant to peekbank json before database
import
}
\examples{
\dontrun{
result <- ds.validate_table(df_table = df_table, table_type = "xy_data", cdi_expected = F, dir_csv = "../processed_data")
result$errors    # blocking issues
result$warnings  # informational warnings
}

}
