% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peekbankr.R
\name{get_administrations}
\alias{get_administrations}
\title{Get administrations}
\usage{
get_administrations(
  age = NULL,
  dataset_id = NULL,
  dataset_name = NULL,
  connection = NULL
)
}
\arguments{
\item{age}{A numeric vector of a single age or a min age and max age
(inclusive), in months}

\item{dataset_id}{An integer vector of one or more dataset ids}

\item{dataset_name}{A character vector of one or more dataset names}

\item{connection}{A connection to the peekbank database}
}
\value{
A `tbl` of Administrations data, filtered down by supplied arguments. If
  `connection` is supplied, the result remains a remote query, otherwise it
  is retrieved into a local tibble.
}
\description{
Get administrations
}
\examples{
get_administrations()
get_administrations(age = c())
get_administrations(dataset_name = "pomper_saffran_2016")
}
