% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peekbankr.R
\name{get_aoi_region_sets}
\alias{get_aoi_region_sets}
\title{Get AOI region sets}
\usage{
get_aoi_region_sets(connection = NULL)
}
\arguments{
\item{connection}{A connection to the peekbank database}
}
\value{
A `tbl` of AOI Region Sets data, filtered down by supplied arguments. If
  `connection` is supplied, the result remains a remote query, otherwise it
  is retrieved into a local tibble.
}
\description{
Get AOI region sets
}
\examples{
get_aoi_region_sets()
}
