% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peekbankr.R
\name{get_aoi_timepoints}
\alias{get_aoi_timepoints}
\title{Get AOI timepoints}
\usage{
get_aoi_timepoints(
  dataset_id = NULL,
  dataset_name = NULL,
  age = NULL,
  connection = NULL
)
}
\arguments{
\item{dataset_id}{An integer vector of one or more dataset ids}

\item{dataset_name}{A character vector of one or more dataset names}

\item{age}{A numeric vector of a single age or a min age and max age
(inclusive), in months}

\item{connection}{A connection to the peekbank database}
}
\value{
A `tbl` of AOI Timepoints data, filtered down by supplied arguments. If
  `connection` is supplied, the result remains a remote query, otherwise it
  is retrieved into a local tibble.
}
\description{
Get AOI timepoints
}
\examples{
get_aoi_timepoints(dataset_name = "pomper_saffran_2016")
get_aoi_timepoints(dataset_name = "pomper_saffran_2016", age = c())
}
