% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peekbankr.R
\name{get_datasets}
\alias{get_datasets}
\title{Get datasets}
\usage{
get_datasets(connection = NULL)
}
\arguments{
\item{connection}{A connection to the peekbank database}
}
\value{
A `tbl` of Datasets data, filtered down by supplied arguments. If
  `connection` is supplied, the result remains a remote query, otherwise it
  is retrieved into a local tibble.
}
\description{
Get datasets
}
\examples{
get_datasets()
}
