% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peekbankr.R
\name{get_readmes}
\alias{get_readmes}
\title{Download dataset README files from OSF to a temporary folder}
\usage{
get_readmes(datasets = c())
}
\arguments{
\item{datasets}{Character vector of dataset names. If empty (default),
downloads READMEs for all datasets.}
}
\value{
No return value, called for side effects. README files are saved to
  a temporary directory and the path is printed via message.
}
\description{
Downloads README files for Peekbank datasets from OSF. Note that READMEs
always reflect the latest version of the dataset on OSF.
}
\examples{
\dontrun{
get_readmes()
get_readmes(datasets = c("pomper_saffran_2016"))
}
}
