% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peekbankr.R
\name{get_sql_query}
\alias{get_sql_query}
\title{Run a SQL Query script on the Peekbank database}
\usage{
get_sql_query(sql_query_string, connection = NULL)
}
\arguments{
\item{sql_query_string}{A valid sql query string character}

\item{connection}{A connection to the Peekbank database}
}
\value{
The database after calling the supplied SQL query
}
\description{
Run a SQL Query script on the Peekbank database
}
\examples{
\dontrun{
get_sql_query("SELECT * FROM datasets")
}
}
