% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peekbankr.R
\name{get_stimuli}
\alias{get_stimuli}
\title{Get stimuli}
\usage{
get_stimuli(dataset_id = NULL, dataset_name = NULL, connection = NULL)
}
\arguments{
\item{dataset_id}{An integer vector of one or more dataset ids}

\item{dataset_name}{A character vector of one or more dataset names}

\item{connection}{A connection to the peekbank database}
}
\value{
A `tbl` of Stimuli data, filtered down by supplied arguments. If
  `connection` is supplied, the result remains a remote query, otherwise it
  is retrieved into a local tibble.
}
\description{
Get stimuli
}
\examples{
\dontrun{
get_stimuli()
get_stimuli(dataset_name = "pomper_saffran_2016")
}
}
