% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peekbankr.R
\name{get_subjects}
\alias{get_subjects}
\title{Get subjects}
\usage{
get_subjects(age = NULL, connection = NULL)
}
\arguments{
\item{age}{A numeric vector of a single age or a min age and max age
(inclusive), in months}

\item{connection}{A connection to the peekbank database}
}
\value{
A `tbl` of Subjects data, filtered down by supplied arguments. If
  `connection` is supplied, the result remains a remote query, otherwise it
  is retrieved into a local tibble.
}
\description{
Get subjects
}
\examples{
get_subjects()
get_subjects(age = c())
}
