% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peekbankr.R
\name{get_subjects}
\alias{get_subjects}
\title{Get subjects}
\usage{
get_subjects(connection = NULL)
}
\arguments{
\item{connection}{A connection to the peekbank database}
}
\value{
A `tbl` of Subjects data. Note that Subjects is a table used to link
  longitudinal Administrations, which is the primary table you probably want.
  If `connection` is supplied, the result remains a remote query, otherwise
  it is retrieved into a local tibble.
}
\description{
Get subjects
}
\examples{
\dontrun{
get_subjects()
}
}
