% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peekbankr.R
\name{get_trial_types}
\alias{get_trial_types}
\title{Get trial types}
\usage{
get_trial_types(dataset_id = NULL, dataset_name = NULL, connection = NULL)
}
\arguments{
\item{dataset_id}{An integer vector of one or more dataset ids}

\item{dataset_name}{A character vector of one or more dataset names}

\item{connection}{A connection to the peekbank database}
}
\value{
A `tbl` of Trial Types data, filtered down by supplied arguments. If
  `connection` is supplied, the result remains a remote query, otherwise it
  is retrieved into a local tibble.
}
\description{
Get trial types
}
\examples{
get_trial_types()
get_trial_types(dataset_id = 0)
get_trial_types(dataset_name = "pomper_saffran_2016")
}
