% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peekbankr.R
\name{get_trials}
\alias{get_trials}
\title{Get trials}
\usage{
get_trials(dataset_id = NULL, dataset_name = NULL, connection = NULL)
}
\arguments{
\item{dataset_id}{An integer vector of one or more dataset ids}

\item{dataset_name}{A character vector of one or more dataset names}

\item{connection}{A connection to the peekbank database}
}
\value{
A `tbl` of Trials data, filtered down by supplied arguments. If
  `connection` is supplied, the result remains a remote query, otherwise it
  is retrieved into a local tibble.
}
\description{
Get trials
}
\examples{
\dontrun{
get_trials()
get_trials(dataset_name = "pomper_saffran_2016")
}
}
