% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peekbankr.R
\name{get_trials}
\alias{get_trials}
\title{Get trials}
\usage{
get_trials(connection = NULL)
}
\arguments{
\item{connection}{A connection to the peekbank database}
}
\value{
A `tbl` of Trials data, filtered down by supplied arguments. If
  `connection` is supplied, the result remains a remote query, otherwise it
  is retrieved into a local tibble.
}
\description{
Get trials
}
\examples{
get_trials()
}
