% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peekbankr.R
\name{get_xy_timepoints}
\alias{get_xy_timepoints}
\title{Get XY timepoints}
\usage{
get_xy_timepoints(
  dataset_id = NULL,
  dataset_name = NULL,
  age = NULL,
  connection = NULL
)
}
\arguments{
\item{dataset_id}{An integer vector of one or more dataset ids}

\item{dataset_name}{A character vector of one or more dataset names}

\item{age}{A numeric vector of a single age or a min age and max age
(inclusive), in months}

\item{connection}{A connection to the peekbank database}
}
\value{
A `tbl` of XY timepoints data, filtered down by supplied arguments.
  If `connection` is supplied, the result remains a remote query, otherwise
  it is retrieved into a local tibble.
}
\description{
Get XY timepoints
}
\examples{
\dontrun{
get_xy_timepoints(dataset_name = "reflook_v4")
}
}
