% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdi_processing.R
\name{populate_cdi_percentiles}
\alias{populate_cdi_percentiles}
\title{Populate the provided cdi data with percentile values for that specific age, instrument_type, measure and language. Loosely based on the work from this repo https://github.com/kachergis/cdi-percentiles/tree/main by George Kachergis and Jess Mankewitz with advice from Virginia Marchman.}
\usage{
populate_cdi_percentiles(subjects_table)
}
\arguments{
\item{subjects_table}{a subjects table with unnested cdi data, needs columns "subject_id", "language", "instrument_type", "age", "sex", "measure", "rawscore"}
}
\value{
the input table with added columns containing the reference age used, the reference year used, and both gender specific and general percentile values for the cdi score
}
\description{
Populate the provided cdi data with percentile values for that specific age, instrument_type, measure and language. Loosely based on the work from this repo https://github.com/kachergis/cdi-percentiles/tree/main by George Kachergis and Jess Mankewitz with advice from Virginia Marchman.
}
\examples{
\donttest{
full_cdi_data <- all_subjects \%>\%
  unnest(subject_aux_data) \%>\%
  filter(!is.na(cdi_responses)) \%>\%
  unnest(cdi_responses) \%>\%
  peekbankr::cleanup_cdi_data() \%>\%
  peekbankr::populate_cdi_percentiles()
}
}
